requirements.txt
streamlit
ultralytics>=8.3.0
opencv-python-headless
Pillow
google-generativeai>=0.2.3
python-dotenv